Pantalla de Inscripciones y agregar/editar usuario
Hubo que hacer arreglos para poder agregar grupos, editar los permisos y
mostrar que usuario esta conectado en el sistema. Ahora hay que
continuar, mostrar un form de inscripcion en una oferta de curso, y
permitir que el estudiante pueda tomar una evaluacion que se haya
ingresado en un curso. Cuando tenga eso, voy a poder mostrar la
pregunta!!!

+++ src/main/java/gamification/service/impl/AuthenticationUserDetails.java
@@ -17,7 +17,7 @@ public class AuthenticationUserDetails implements org.springframework.security.c
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private Long id;
+	private int id;
     private final String login;
     private final String passwordHash;
     private final boolean enabled;
@@ -87,7 +87,7 @@ public class AuthenticationUserDetails implements org.springframework.security.c
         return login;
     }
 
-	public Long getId() {
+	public int getId() {
 		log.trace("Estoy en getId");
 		return id;
 	}
